gcp-fishdisease/requirements.txt
tensorflow==2.8.0
google-cloud-storage==1.16.1
pillow
protobuf==3.20.0